tutorial/requirements.txt
Django
djangorestframework
paramiko